added translations for messages from openai_helper
